% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_batch_file.R
\name{load_batch_file}
\alias{load_batch_file}
\title{Parse a batch file and return the data as dataframe.}
\usage{
load_batch_file(file)
}
\arguments{
\item{file}{The name of the xlsx batch file.}
}
\value{
A dataframe with the contents of the batch file.
}
\description{
Parse a batch file and return the data as dataframe.
}
\author{
Daniel Montemayor, \email{montemayord2@uthscsa.edu}
}
